requirements.txt
Django==3.2.7
djangorestframework==3.12.4
pygments==2.10.0
django_heroku==0.3.1
gunicorn 
dj-database-url 
whitenoise 
psycopg2-binary